requirements.txt
fastmcp==0.1.0
pydantic==2.6.1
python-dateutil==2.8.2
python-docx>=0.8.11
PyPDF2>=3.0.0
nltk>=3.8.1
spacy>=3.7.2
transformers>=4.36.0
torch>=2.1.0
python-magic>=0.4.27 